add rest module

+++ common-jpa/src/main/java/ru/relex/dao/AppPhotoDAO.java
@@ -3,5 +3,5 @@ package ru.relex.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.relex.entity.AppPhoto;
 
-public interface AppPhotoDAO extends JpaRepository<AppPhoto, Integer> {
+public interface AppPhotoDAO extends JpaRepository<AppPhoto, Long> {
 }
